requirements.txt
streamlit>=1.28.0
langchain-google-genai>=1.0.0
langgraph>=0.0.30
langchain-core>=0.1.0
google-generativeai>=0.3.0
streamlit-lottie>=0.0.5
requests>=2.31.0
pypdf>=3.17.0
langchain-community>=0.0.20
chromadb>=0.4.22
langchain-text-splitters>=0.0.1
